+++ app/src/main/java/com/example/view/adapter/SubCategoryAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.adapter
+package com.example.view.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
