+++ app/src/main/java/com/example/view/fragment/checkout/SummaryFragment.kt
@@ -1,4 +1,4 @@
-package com.example.fragment.checkout
+package com.example.view.fragment.checkout
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,26 +6,23 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.viewpager2.widget.ViewPager2
-import com.example.adapter.CheckoutItemAdapter
+import com.example.view.adapter.CheckoutItemAdapter
 import com.example.e_commereceapplication.R
-import com.example.e_commereceapplication.databinding.FragmentCartItemsBinding
+import com.example.e_commereceapplication.databinding.FragmentSummaryBinding
 import com.example.model.local.DbHandler
 import com.example.model.local.dao.ProductDao
 
+class SummaryFragment : Fragment() {
 
-class CartItemsFragment : Fragment() {
-
-    private lateinit var binding: FragmentCartItemsBinding
+    private lateinit var binding: FragmentSummaryBinding
     private lateinit var dbHandler: DbHandler
     private lateinit var productDao: ProductDao
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         initDao()
-        binding = FragmentCartItemsBinding.inflate(inflater, container, false)
+        binding = FragmentSummaryBinding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -39,16 +36,13 @@ class CartItemsFragment : Fragment() {
         binding.apply {
             rcProducts.adapter = adapter
             rcProducts.layoutManager = layoutManager
-
-            txtTotal.text = "$ ${productDao.calculateTotalPriceInCart()}"
-
-            binding.btnNext.setOnClickListener {
-                val viewPager = requireActivity().findViewById<ViewPager2>(R.id.viewPager2)
-                val currentItem = viewPager.currentItem
-                viewPager.setCurrentItem(currentItem + 1, true)
-            }
-
         }
+
+
+        binding.apply {
+            txtTotalBillHolder.text = "$ ${productDao.calculateTotalPriceInCart()}"
+        }
+
     }
 
     private fun initDao() {
