+++ app/src/main/res/layout/fragment_delivery.xml
@@ -15,13 +15,48 @@
         app:layout_constraintStart_toEndOf="@id/vertical_left"
         app:layout_constraintTop_toBottomOf="@id/horizontal_up" />
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rc_addresses"
+    <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        tools:listitem="@layout/address_item"
-        app:layout_constraintBottom_toTopOf="@id/btn_add_address"
-        app:layout_constraintTop_toBottomOf="@id/txt_selectAddy" />
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toBottomOf="@id/txt_selectAddy"
+        android:padding="16dp">
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="300dp"
+            android:padding="10dp"
+            android:layout_gravity="center_horizontal"
+            android:layout_height="wrap_content">
+
+            <androidx.appcompat.widget.AppCompatTextView
+                android:id="@+id/txt_title"
+                style="@style/NavigationViewTextAppearance"
+                android:textColor="@color/background_cardview"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/home_addy"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintStart_toStartOf="parent"/>
+
+            <androidx.appcompat.widget.AppCompatTextView
+                android:id="@+id/txt_address"
+                android:layout_width="@dimen/_0dp"
+                android:layout_height="wrap_content"
+                android:text="@string/dummy_addy"
+                android:textSize="@dimen/_17sp"
+                app:layout_constraintEnd_toStartOf="@id/selected_address"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/txt_title"/>
+
+            <RadioButton
+                android:id="@+id/selected_address"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"/>
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+    </androidx.cardview.widget.CardView>
 
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/btn_add_address"
